# swagger.yaml
openapi: 3.0.0
info:
    title: DB Sync API
    version: 1.0.0
    description: APIs for cardano DB Sync
paths:
...
    /api/drep/{id}/active-delegators:
        get:
            summary: Get active delegators for a DRep
            description: Fetch the list of active delegators associated with the specified DRep ID.
            tags:
                - DRep Delegation
            parameters:
                - name: id
                  in: path
                  required: true
                  description: The DRep ID in hex or Bech32 format.
                  schema:
                      type: string
                      example: 'drep15fy2nszna039pdregpryhzahnwcr2rred4nx2qaec7k07nvcq26'
                - in: query
                  name: page
                  schema:
                      type: number
                      default: 1
                  required: false
                - in: query
                  name: size
                  schema:
                      type: number
                      default: 10
                  required: false
                - name: balance
                  in: query
                  required: false
                  description: The active balance of the delegator.
                  schema:
                      type: boolean
                      example: false
            responses:
                '200':
                    description: A list of active delegators.
                    content:
                    application/json:
                        schema:
                            type: array
                            items:
                                type: object
                                properties:
                                    delegatorStake:
                                        type: string
                                        description: The stake address of the delegator.
                                    balance:
                                        type: object
                                        properties:
                                            utxoBalance:
                                                type: string
                                                description: The UTXO balance of the delegator.
                                            rewardBalance:
                                                type: string
                                                description: The reward balance of the delegator.
                                            rewardRestBalance:
                                                type: string
                                                description: The reward rest balance of the delegator.
                                    delegatedAt:
                                        type: object
                                        properties:
                                            txId:
                                                type: string
                                                description: The transaction ID of the delegation.
                                            epochNo:
                                                type: integer
                                                description: The epoch number of the delegation.
                                            time:
                                                type: string
                                                format: date-time
                                                description: The time of the delegation.
                '400':
                    description: Invalid DRep ID.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                message:
                                    type: string
                                    example: Provide a valid Drep ID
                '500':
                    description: Internal server error.
...
    /api/drep/{id}:
        get:
            summary: Get DRep details
            description: Fetches DRep details including registration status, metadata, voting power, and transaction hashes for the provided address.
            tags:
                - DRep Information
            operationId: getDrepDetails
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Hex-encoded address or Bech32 address of DRep .
                  schema:
                      type: string
                      example: 'drep15fy2nszna039pdregpryhzahnwcr2rred4nx2qaec7k07nvcq26'
            responses:
                '200':
                    description: Successful response with DRep and SoleVoter details.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    hasScript:
                                        type: boolean
                                        description: whether the drep has script or not
                                    isRegisteredAsDRep:
                                        type: boolean
                                        description: Whether the address is currently registered as a DRep.
                                    wasRegisteredAsDRep:
                                        type: boolean
                                        description: Whether the address was previously registered as a DRep.
                                    isRegisteredAsSoleVoter:
                                        type: boolean
                                        description: Whether the address is currently registered as a Sole Voter.
                                    wasRegisteredAsSoleVoter:
                                        type: boolean
                                        description: Whether the address was previously registered as a Sole Voter.
                                    deposit:
                                        type: integer
                                        description: The current deposit value.
                                    url:
                                        type: string
                                        description: The metadata URL.
                                    dataHash:
                                        type: string
                                        description: Hash of the metadata.
                                    votingPower:
                                        type: integer
                                        description: The current voting power.
                                    dRepRegisterTxHash:
                                        type: string
                                        description: Transaction hash for DRep registration.
                                    dRepRetireTxHash:
                                        type: string
                                        description: Transaction hash for DRep retirement.
                                    soleVoterRegisterTxHash:
                                        type: string
                                        description: Transaction hash for Sole Voter registration.
                                    soleVoterRetireTxHash:
                                        type: string
                                        description: Transaction hash for Sole Voter retirement.

                '400':
                    description: Bad Request. Address is invalid or missing.
                '500':
                    description: Internal Server Error. Something went wrong with the query or database.
    /api/drep/{id}/vote:
        get:
            summary: Get voting details by DRep ID
            description: Retrieve voting details for a specific DRep ID.
            tags:
                - DRep Information
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Hex-encoded address or Bech32 address of DRep .
                  schema:
                      type: string
                      example: 'drep15fy2nszna039pdregpryhzahnwcr2rred4nx2qaec7k07nvcq26'
            responses:
                '200':
                    description: Successfully retrieved voting details of given Drep
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    govActionId:
                                        type: string
                                        description: Unique identifier for the government action, combining hash and index.
                                        example: '8af0c0c97f532ebd72df62cf798ac19fc2ec952d411ece7de89#0'
                                    title:
                                        type: string
                                        nullable: true
                                        description: Title of the government action.
                                        example: 'Example Title'
                                    voteType:
                                        type: string
                                        description: Type of vote.
                                        example: 'Abstain'
                                    voteAnchorUrl:
                                        type: string
                                        format: uri
                                        description: URL of the vote anchor.
                                        example: 'https://example.com/aMBitupUUJucUCAg9sxVyHe4jjk41dzst9J'
                                    voteAnchorHash:
                                        type: string
                                        description: Hash of the vote anchor data.
                                        example: 'd6086a79267ab5fb7b56c8ab77f1e0dff2c3'
                                    epochNo:
                                        type: integer
                                        description: Epoch number associated with the government action.
                                        example: 762
                                    time:
                                        type: string
                                        format: date-time
                                        description: Timestamp of the action.
                                        example: '2024-11-25T07:23:50'
                                    voteTxHash:
                                        type: string
                                        description: Transaction hash of the vote.
                                        example: '8d09be316ced784c1dd5bbcf7dccc0d63ea8de44b6e'
                                    govActionType:
                                        type: string
                                        description: Type of government action.
                                        example: 'NoConfidence'
                '404':
                    description: DRep ID not found
                '500':
                    description: Internal server error
    /api/drep/{id}/delegation:
        get:
            summary: Get delegation details by DRep ID
            description: Retrieve delegation details for a specific DRep ID.
            tags:
                - DRep Delegation
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Hex-encoded address or Bech32 address of DRep.
                  schema:
                      type: string
                      example: 'drep15fy2nszna039pdregpryhzahnwcr2rred4nx2qaec7k07nvcq26'
                - in: query
                  name: page
                  schema:
                      type: number
                      default: 1
                  required: false
                - in: query
                  name: size
                  schema:
                      type: number
                      default: 10
                  required: false
            responses:
                '200':
                    description: A list of stake delegations.
                    content:
                        application/json:
                            schema:
                            type: array
                            items:
                                stakeAddress:
                                    type: string
                                    example: 'stake_test17pmm4fm6aahntj59y7p2xkmhkppp2yc3qlejgln997dcy4scqeut3'
                                action:
                                    type: string
                                    example: 'joined'
                                txId:
                                    type: string
                                    example: '08e5daeb18137efd8ad3ddbda6347f490a3efa8f839da3e2a18836ba0ff1bba8'
                                epochNo:
                                    type: integer
                                    example: 818
                                time:
                                    type: string
                                    example: '2025-01-20T08:35:02'
                '404':
                    description: DRep ID not found
                '500':
                    description: Internal server error
    /api/drep/{id}/registration:
        get:
            summary: Get registration details by DRep ID
            description: Retrieve registration details for a specific DRep ID.
            tags:
                - DRep Information
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Hex-encoded address or Bech32 address of DRep .
                  schema:
                      type: string
                      example: 'drep15fy2nszna039pdregpryhzahnwcr2rred4nx2qaec7k07nvcq26'
            responses:
                '200':
                    description: Successfully retrieved voting details of given Drep
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    url:
                                        type: string
                                        example: 'https://xyz.com/drep-id/data/main/preview/test'
                                    deposit:
                                        type: integer
                                        nullable: true
                                        example: null
                                    drepName:
                                        type: string
                                        example: 'name'
                                    time:
                                        type: string
                                        format: date-time
                                        example: '2024-12-12T08:11:03.000Z'
                '404':
                    description: DRep ID not found
                '500':
                    description: Internal server error
...
    /api/drep:
        get:
            summary: Fetch DRep list
            description: Fetches the drep list.
            tags:
                - DRep Information
            parameters:
                - in: query
                  name: page
                  schema:
                      type: number
                      default: 1
                  required: false
                - in: query
                  name: sort
                  schema:
                      type: string
                      enum: [VotingPower, RegistrationDate]
                - in: query
                  name: status
                  schema:
                      type: string
                      enum: [Active, Inactive, Retired]
                - in: query
                  name: size
                  schema:
                      type: number
                      default: 10
                  required: false
                - in: query
                  name: search
                  description: Search by dRepId, dRepView or dRep Name
                  schema:
                      type: string
                      default: ''
                  required: false
            responses:
                200:
                    description: 'A paginated list of DRep details.'
                    content:
                        application/json:
                            schema:
                                type: 'object'
                                properties:
                                    items:
                                        type: 'array'
                                        items:
                                            type: 'object'
                                            properties:
                                                drepId:
                                                    type: 'string'
                                                    example: 'string'
                                                    description: 'DRep ID in hexadecimal format.'
                                                view:
                                                    type: 'string'
                                                    example: 'string'
                                                    description: "DRep's view identifier."
                                                url:
                                                    type: 'string'
                                                    format: 'uri'
                                                    example: 'string'
                                                    description: "URL to the DRep's metadata."
                                                metadataHash:
                                                    type: 'string'
                                                    example: 'string'
                                                    description: 'Hash of the metadata.'
                                                deposit:
                                                    type: 'integer'
                                                    format: 'int64'
                                                    example: 0
                                                    description: 'Deposit amount.'
                                                votingPower:
                                                    type: 'integer'
                                                    format: 'int64'
                                                    example: 0
                                                    description: "DRep's amount."
                                                active:
                                                    type: 'boolean'
                                                    example: false
                                                    description: 'Indicates if the DRep is active.'
                                                latestTxHash:
                                                    type: 'string'
                                                    example: 'string'
                                                    description: 'Hash of the latest transaction.'
                                                latestRegistrationDate:
                                                    type: 'string'
                                                    format: 'date-time'
                                                    example: '0000-00-00T00:00:00'
                                                    description: 'Date and time of the latest registration.'
                                    page:
                                        type: 'integer'
                                        example: 0
                                        description: 'Current page number.'
                                    size:
                                        type: 'integer'
                                        example: 0
                                        description: 'Number of items returned per page.'
                                    total:
                                        type: 'integer'
                                        example: 0
                                        description: 'Total number of items'
                '400':
                    description: Invalid address provided
                '500':
                    description: Internal server error
...